Add Invoke Callback to `webview_alloc`
Allow setting more of the webview state up in the initial
`webview_alloc` call. This seems easier than adding a bunch of
`webview_set_xxx` methods to the API. Seen as this is intended for ffi
callers the state can be buffered on the other side of the interface
rather than having to worry about making this API more ergonomic.