Windows: Hide window until run
Because the window is shown as soon as the webview gets created, it is possible, that the end users see a 'flickering' window. For example, if the developer sets the size of the window, first a window is shown with default sizes and no content, then it gets resized (flickering).

This commit creates a 'hidden' window on initialization and it gets shown on calling the run method. In between of the initialization and running the webview the change of configuration parameters, through calling methods like set_title and set_size, will only affect the hidden window and it will not occur as 'flickering' to the user.