Change Edge/Chromium user data folder
Previously the code passed a nullptr to the creation of the Edge/Chromium Webview. This defaults the webview to create a user data folder in the folder of the calling executable. In case of an installed application, the executing user may not have the permissions to do this. This commit changes the default user data folder to "%APPDATA%/[EXECUTABLE_NAME]". Fixes #362