Use C compiler for C example
This documents @petabyt's approach for compiling webview with
a C++ compiler and the C program with a C compiler, then linking
them together. This requires setting WEBVIEW_HEADER in C to tell
gcc to only use the header and ignore the rest of webview.h, which
is in C++, so gcc can't compile it.

The explanation is much longer than before, because there are now
more steps. To simplify it, we first set an environment variable
for platform-specific C++ flags and keep the build process the
same for Linux, macOS, and Linux.

This approach is better than using a C++ compiler to compile C++,
since C and C++ have some discrepancies. It also has the added
bonus of faster incremental and initial compilation.

The results are identical as before, with a webview-example built
in the working directory.