Add C example to main repo
This adds the C example found at webview/webview_c to the main
repository, and configures the Unix build scripts to test it in CI.
This will ensure that the C example is constantly tested as well as
the C++ one.

It currently does not add the C example to the Windows tests. That
can either be added to this PR or separately if desired.