Improve convenience for library consumers
Build output, tests and examples were moved tidy up a bit. Go-related files were not moved because of new problems.

The (sh) build script was cleaned up slightly because of the increasing complexity.

Facilitate different types of libraries: header-only (default), shared and static. For now, Windows/MSVC builds have been left more or less as-is.

New options for compiling and using the library: WEBVIEW_BUILDING for use when building the library; WEBVIEW_SHARED for use when building or consuming the library as a shared library; WEBVIEW_STATIC for use when building or consuming the library as a static library; WEBVIEW_INCLUDE_IMPLEMENTATION is used internally but can be used to force inclusion of the library implementation (use together with WEBVIEW_API) when compiling in C++ mode.

WEBVIEW_HEADER can still be used to opt out of including the library implementation when consuming the library.

The library implementation can no longer be included in C mode.

Finally, this work solves a problem when using the library as a header-only library in C++ mode because of multiple definitions which are now marked as inline only in this specific case.