Fix resource leaks causing hang on Windows
Reference counting is now implemented for webview2_com_handler so that
it can be released properly. The cloned pointers to ICoreWebView2Controller
and ICoreWebView2 are now being released and it is more clear when the
pointers are cloned.

These actions fix an issue where WebView2 initialization can hang for a
variable amount of time and cause tests to time out.

With the old code on my local environment, tests that use WebView2
oftentimes hang for about a minute. We have also seen how the CI job
tends to hang, likely for the same reason.

I do not think that it has anything to do with the tests themselves but
rather the right sequence of events paired with leaks in a short period
of time.

The hang happens after a successful call to
ICoreWebView2Environment::CreateCoreWebView2Controller
and before getting the webview controller in
ICoreWebView2CreateCoreWebView2ControllerCompletedHandler. No browser
instances are created while it is hanging. After some time the browser
instances (msedgewebview2.exe) suddenly appear, we get the controller
and can move on.