Use correctly sized window icon on Windows
Fixes window icon by loading a correctly sized icon and setting it as the
hIcon member of WNDCLASSEXW. Previously a small icon was always loaded and
set for the hIcon member.

If we do not have a small icon then we do not need to specify the hIconSm
member as the system will search hIcon for an appropriately sized icon.