Fix Windows CI build (#785)
* Use cmd shell explicitly for Windows CI job

This somehow fixes the CI job apparently succeeding even though the script returned non-zero.

* Use the correct name for the WebView2Loader library for Go

* Use CGO_CXXFLAGS instead of CGO_CPPFLAGS in Windows build script

* Set PATH environment variable for Go tests

* Use setlocal in Windows build script

Keeps the shell cleaner by discarding environment variables when the script exits.

* Fix double backslash in paths in Windows build script

* Document argument quoting issue for CGO_CXXFLAGS/CGO_LDFLAGS

* Fix Windows build script exit command

It is "exit /b", not "exit \b". Fixes the command terminating the whole terminal.

* Move Go environment setup in Windows build script

* Add MinGW-w64 to PATH

GitHub Actions somehow does not have this directory in the PATH environment variable.

This fix allows programs built with MinGW-w64 to find runtime libraries and is more explicit about which version of MinGW-w64 will be used for compilation.