Add COM library initialization (#761)
The WebView2 COM API states that the application is expected to call
CoInitializeEx before CreateCoreWebView2EnvironmentWithOptions:

    CoInitializeEx(nullptr, COINIT_APARTMENTTHREADED);

WebView2 cannot function without someone calling CoInitializeEx and
this webview library cannot function on Windows without WebView2.

Although it may appear to work without calling CoInitializeEx, it is
eventually called with COINIT_APARTMENTTHREADED somewhere down the call
stack when we call ShowWindow and pass in SW_SHOW. This is however not
always the case because when we pass in something like SW_HIDE to
ShowWindow then CoInitializeEx will no longer be called from ShowWindow.

If ShowWindow is called with SW_HIDE or if it is not called at all then
SetFocus will cause CoInitializeEx to be called instead. If SetFocus is
then not called (why would we do so on a hidden window?) then
CoInitializeEx is no longer called from anywhere.

With nothing else calling CoInitializeEx with COINIT_APARTMENTTHREADED,
WebView2 initialization fails.

This shows that we have been relying on a side effect.

A simple way for us to have a more consistent behavior is to call
CoInitializeEx in the core library automatically. There is no harm in
calling CoInitializeEx multiple times with the same concurrency model.
This will also allow bindings to not have to concern themselves with
COM library initialization.