Fix app termination
NSApp is not a class but a global variable of the type NSApplication.
Attempting to send a termination message to that "class" does nothing.
Sending the same message to the shared NSApplication instance instead
will terminate the application.

It is not necessary to both close the window and terminate the application,
and attempting to close a window that is not yet visible does not terminate
the application.