Fix crash when WebView2 runtime not installed
This fixes an issue on Windows where webview construction would cause the
program to crash when the WebView2 runtime was not installed.
webview_create now returns null this case.

CreateCoreWebView2EnvironmentWithOptions would return
HRESULT_FROM_WIN32(ERROR_FILE_NOT_FOUND) (0x80070002); the value
returned by embed() (false) would not be checked in the constructor
of win32_edge_engine; it would then crash when attempting to call
MoveFocus() via m_controller which was nullptr.

This fix does not add a check for embed() but instead checks much earlier
if the WebView2 runtime is available, avoiding creating the window just
to destroy it again.

We should still check the return value of embed() but this simple fix
requires less testing and can therefore be merged sooner.