Rearrange macOS initialization code
Objective-C projects created with Xcode normally call NSApplicationMain()
in main(). It creates the application, loads the main nib file and starts
the main run loop.

The implementation of NSApplicationMain is similar to the following:

    void NSApplicationMain(int argc, char *argv[]) {
        [NSApplication sharedApplication];
        [NSBundle loadNibNamed:@"myMain" owner:NSApp];
        [NSApp run];
    }

NSApplication posts NSApplicationDidFinishLaunchingNotification to indicate
that the app has completed launching and is ready to run. This happens
when invoking [NSApp run] and lets us perform additional initialization in
the applicationDidFinishLaunching: method of the app delegate.