Expect applicationDidFinishLaunching only when a window was not passed in
On macOS, initializing the library with an existing window likely means
that the user wants to be responsible for setting up the application, and
likely means that the user intends to manage the application lifecycle.
If the library waits for the "applicationDidFinishLaunching:" message in
this case then it may never receive the it.

Since the library still replaces any existing app delegate with its own,
this change is consistent with behavior that existed in the library before
it began expecting the "applicationDidFinishLaunching:" message.