Redo clang-tidy config (#767)
The main purpose of this work is not to fix all issues in the code but to make a new, cleaner and more lax ruleset so that we can gradually fix issues and remove suppressions. The existing ruleset produced thousands of lines with warnings cluttering the build log, causing real issues buried within to likely go unnoticed.