Custom WebView2Loader implementation (#783)
A minimalistic reimplementation of the WebView2 loader to be used when the
official loader is not present.

Bundling WebView2Loader.dll is now optional when using the built-in loader
implementation. Explicit linking of WebView2Loader.dll is now enabled by
default when the built-in loader is enabled to avoid the need for import
libraries (*.lib).

Compile-time options can be used to enable or disable the built-in loader
and explicit linking of the DLL: WEBVIEW_MSWEBVIEW2_BUILTIN_IMPL,
WEBVIEW_MSWEBVIEW2_EXPLICIT_LINK.

See the readme file for more details.