Conditionally guard usage of new API
Adds a runtime check to avoid crashing the program on OS versions that
do not support the setInspectable: selector.

To add some more information, this combined work makes WKWebView
inspectable via Safari on OS versions that disable the feature by
default (macOS 13.3 and later) and support enabling it.

According to Apple, the behavior on older OS versions is for content
to always be inspectable in "debug builds".

Testing shows that this is true for macOS 12.6 but somehow not 10.15.