Disable status bar for MS WebView2
With MS WebView2 on Windows, hovering the mouse cursor over links would
display a status bar in the bottom-left corner of the top-level window.

If the mouse cursor was located in the corner of the window then the
status bar would move from inside the window to outside of the window.

The status bar is now disabled to align the behavior with other
backends.