Explicitly set GHA image and compiler versions
The ubuntu-22.04 GHA image appears to have been updated in some way
since the last commit in master which has broken future CI builds.

The issue appears to be that Clang (Tidy) 14 uses GCC 13 headers and
somehow that appears to be an issue as shown below:

    /usr/lib/gcc/x86_64-linux-gnu/13/../../../../include/c++/13/chrono:2320:48: error: call to consteval function 'std::chrono::hh_mm_ss::_S_fractional_width' is not a constant expression [clang-diagnostic-error]
            static constexpr unsigned fractional_width = {_S_fractional_width()};
                                                        ^
    /usr/lib/gcc/x86_64-linux-gnu/13/../../../../include/c++/13/chrono:2320:48: note: undefined function '_S_fractional_width' cannot be used in a constant expression
    /usr/lib/gcc/x86_64-linux-gnu/13/../../../../include/c++/13/chrono:2275:2: note: declared here
            _S_fractional_width()
            ^

This work "fixes" (works around) this issue by pinning Clang tools to a
a more recent version.

While making this change for Clang tools, I thought it would sense to
do the same for GCC, and also pin the GHA image to ubuntu-22.04 instead
of ubuntu-latest.

Old defaults:
* GCC: 11
* Clang: 14

New defaults:
* GCC: 12
* Clang: 15