Workaround for blank window with WebKit/DMA-BUF/NVIDIA/X11
WebKit 2.42 along with X11 and NVIDIA GPU driver triggers bug:
https://bugs.webkit.org/show_bug.cgi?id=261874

The bug causes a blank window. Instead of waiting for the problem
to be fixed, this work attempts to check for these conditions before
setting the environment variable WEBKIT_DISABLE_DMABUF_RENDERER to 1.

If the variables WEBKIT_DISABLE_DMABUF_RENDERER or
WEBKIT_DISABLE_COMPOSITING_MODE have already been set then the
workaround is not applied.