Allow passing HWND to webview_create (Win32)
The function originally took a pointer to HWND, but it's more natural
for a caller to pass an HWND which by itself is a pointer.

Support for passing a pointer to HWND is retained by checking whether
the supplied value is an existing window.

Documentation was reworded a bit for clarity.