Set focus to webview widget when owned window activates (Win32) (#1073)
Previously, the webview widget would lose focus when the top-level
window went into the background and then into the foreground.

This behavior would break the workflow e.g. during alt-tabbing by
requiring the user to click with the mouse inside the webview widget
in order to regain focus.

The behavior on Linux is that focus is automatically regained.