Embed WebView2 into a child window (Win32) (#1074)
Linux/GTK and macOS/Cocoa both support automatic UI layouting and can
therefore automatically resize the webview widget when embedded into a
non-owned window, but Windows/Win32 has no automatic layout system.
I don't believe that we can non-intrusively determine when the widget
should be resized. An intrusive way could be to replace the non-owned
window's WndProc.

This work creates a UI widget (child window) that WebView2 can render
onto, and this widget is embedded into the top-level window.
If we expose the native handle of this widget then we can let the owner
of the window take responsibility of resizing the widget.