Last window closed custom logic (macOS) (#1085)
* Split window/web view set-up into separate functions (macOS)

* Control stopping the run loop when last window is closed (macOS)

We used to let the system/UI framework automatically handle when to
terminate the app after the last window was closed. We used custom logic
to stop the run loop instead of terminating the whole process.

Now we handle the stopping of the run loop without having to awkwardly
initiate termination just to block it and do something else instead.

Future work can e.g. add a flag that allows the user of the library to
choose whether to "terminate" (stop the run loop) when the last window
is closed.