Quit main loop after processing event queue (GTK)
Aligns behavior with other backends while also fixing issue with
resolve() being called in or after a binding callback and work is put
on the event queue without being cleared or processed until the next
call to run(). The next run() would then attempt to process the work,
but the webview instance could be different.