Simplify WebKitGTK compatibility code
Since webkit_web_view_run_javascript became deprecated in WebKitGTK
2.40, it lead to compile warnings. The remedy was to dynamically import
that function and the modern webkit_web_view_evaluate_javascript,
and call them based on which version of the WebKitGTK library was
loaded.

With the recent changes to add support for WebKitGTK API 6.0 and GTK 4,
some existing functions in those libraries have either been removed,
replaced or had their signatures changed. Compiling in support for all
the different versions will likely be difficult to maintain over time.

Therefore this piece of code was changed from being a runtime decision
to a compile-time decision.