Fix some memory leaks
The WebKitWebView instance appears to still leak when using WebKitGTK
API 6.0 and then terminating the main loop early, but if we were to
spin the main loop some more times then it appears that the instance
will become freed eventually.

There are other apparent leaks though that I don't know how to fix.