Fix clang-format conflict before and after v16
clang-format-16 and later (18) do not agree with clang-format-15 and
earlier (9) on how to format the member initializers in the constructor
of cocoa_wkwebview_engine.

By enabling ConstructorInitializerAllOnOneLineOrOnePerLine, the
initializers are formatted consistently either on a single line or on
separate lines if they can't fit on a single line.

This option was introduced in clang-format 3.7 and is currently
deprecated. The replacement appears to be PackConstructorInitializers,
which was introduced in clang-format-14. The deprecated option is used
here to retain the same clang-format (9) compatibility.