Fix deletion of incomplete type user_script::impl
When compiling in C++23 mode using Clang, compilation
failed in ~unique_ptr() due to user_script::impl being an incomplete type.

Issue was observed with Clang 16 and 18 on Linux/macOS, with C++23 set via
CMAKE_CXX_STANDARD. C++20 and earlier did not yield an error.

The issue was could not be observed with GCC 13.2 and 14.2 on Linux.