Adjustments to amalgamation script with test
Amalgamation script now fails by default when a header can't be found.
Missing headers can be explicitly allowed with a comment:

    #include "xyz.h" // amalgamate(skip)

Inputs can now also be paths to directories in which any file or
directory will be handled as an input. Inputs are handled recursively.

The amalgamated header is now not only generated in CI in order to test
the script itself, but the common use cases for the header are now
tested by building a simple program for different platforms against the
generated header.