C++11 compatibility for Windows (#1260)
This work makes some changes that are necessary for lowering the
minimum required C++ standard on Windows from C++14 to C++11.

I can't recall why cxx_std_14 was set instead of cxx_std_11 for the
MSWebView2::headers CMake target but it may have been under the wrong
belief that it wouldn't work with cxx_std_11. Now it does appear to
work with the MinGW distributions we test with so making this change
seems reasonable to me.

Some code for Windows were using C++14 features and have been replaced
to be compatible with C++11.

While at it, the move constructor and assignment operator of
com_init_wrapper should have been marked as noexcept and has
therefore been marked as such.