Adjustments to handling of long-running tests (#1275)
There has been a problem for some time where the first test that
cold-starts an MS WebView2 instance in a GitHub Actions environment
may be severely delayed and result in a timeout. The delay has been
observed to range from nearly no time to a minute or longer.

While this doesn't solve the root cause, it should at least alleviate
some pain due to timeouts.

This work does the following:

* Separates unit tests from functional tests. They're now in separate
  source files and are compiled into separate executables. This is not
  strictly necessary but organizes tests a bit better.

* The CMake configuration for the test driver now supports setting a
  normal timeout value as well as an additional timeout value that takes
  precedence whenever a test prints a matching string to standard
  output/error.

* Adds a dummy warm-up test that runs first only on Windows with a
  longer timeout using the mechanism described above.