Remove user-defined literal operators
The _cls, _sel and _str UDL operators/suffixes have been removed to
avoid a messy situation.

The UDLs can't cleanly be used without "using namespace", which adds
noise to functions, and when used at namespace scope then we pollute
the namespace. The UDLs are so widely in the cocoa and webkit namespaces
that noise/pollution seems to be unavoidable.

I wanted to move _str from the objc::literals namespace into the
cocoa::literals namespace where it's more appropriate since it
instantiates the NSString class; however, it would've been inaccessible
without "using namespace cocoa::literals".

cocoa::literals could pull in everything from objc::literals, and
webkit::literals could pull in everything from cocoa::literals and
objc::literals. This is not too bad because the literals are used
everywhere within the cocoa and webkit namespaces.

However, that doesn't solve the need for "using namespace x::literals"
within the cocoa and webkit namespaces where functions are defined,
which somewhat defeats the purpose of having the UDLs in a separate
namespace.

One way or another it seems like it'll get messy.

Instead the UDLs have been replaced with the regular functions
objc::get_class(), objc::selector() and NSString_stringWithUTF8String().

This eliminates the whole issue with noise and pollution at the cost of
more verbose code. Browser backend code does however not become more
verbose by this change as it doesn't directly use the UDLs.

While NSString_stringWithUTF8String is certainly more verbose than _str,
it isn't used so frequently that it's an issue.

In conclusion, the UDL operators were removed so that noise and
pollution related to them could be eliminated, even at a cost of
slightly increased code verbosity in locations that do not need to
be touched very often.

Other changes:

* Removed unnecessary C-style casts to "Class" and "id".
  Since the UDLs were replaced and the new functions return the correct
  types unlike the UDLs, no more casting is needed.